+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming Next Release
 
+* Fixed when ancestor is not CIO (class, record or interface) for PasDoc (Michalis)
 * Parsing "unimplemented" directive (Michalis)
 * [Completely new default look: using the latest Bootstrap for modern base style, mobile-friendly, nicer visibility pills, new --css-based-on-bootstrap, show only available sections, backward compatible CSS preserved](https://github.com/pasdoc/pasdoc/discussions/230) (Michalis)
 * [New feature to show source code file/line + link in docs: `--show-source-position`, `--source-url-pattern`, `--source-root`](https://pasdoc.github.io/SourcePosition), see also [discussion thread](https://github.com/pasdoc/pasdoc/discussions/229) (Michalis)